requirements.txt
python-telegram-bot==20.4
requests
beautifulsoup4
nest_asyncio
lxml